Move index to .tsx, features template
