+++ src/types/static.d.ts
@@ -1,0 +1,4 @@
+declare module '@site/*' {
+  const content: any;
+  export default content;
+}
